+++ release_notes.md
@@ -1,5 +1,10 @@
 ## Release Notes 2.12
 
+## Release Notes 2.12.1 2022-05-26
+- ibllib.pipes.training_status: pipeline to compute training status of mice on local servers, new TrainingStatus task (Mayo)
+- Fix swanson regions (Olivier)
+- Swanson to Beryl mapping
+
 ### Release Notes 2.12.0 2022-05-10
 - ibllib.atlas: add the Swanson flatmap backend (Olivier)
 - ibllib.io.extractors: output of task extractions are trial tables, not individual datasets (Miles)